Created UI hierarchy, switched a couple settings
Changed the Main scene to a control node so we can instantiate and deinstantiate gameplay scenes as needed. Switched to Compatibility renderer, we shouldn't need the 3D lighting involved with that. Switched to 1280 by 720 screen size. Switched to default full screen. Set up the base scenes for Music, SFX, and Data Storage Nodes(to be written later).